work on lesson2 completed
